python-ai-server/main.py
"""
Python AI Server - MONSTA Trading
실제 AI 모델 실행을 담당하는 FastAPI 서버
"""

from fastapi import FastAPI, WebSocket, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Dict, Optional
import numpy as np
import pandas as pd
from datetime import datetime, timedelta
import asyncio
import json
import redis
import os
from dotenv import load_dotenv

# AI 모델 임포트 (추후 구현)
# from models.lstm import LSTMPredictor
# from models.xgboost import XGBoostPredictor
# from models.arima import ARIMAPredictor

load_dotenv()

app = FastAPI(title="MONSTA AI Server", version="1.0.0")

# CORS 설정
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Redis 연결 (캐싱용) - 옵셔널
redis_client = None
try:
    redis_client = redis.Redis(
        host='localhost',
        port=6379,
        db=0,
        decode_responses=True
    )
    redis_client.ping()
    print("Redis connected")
except:
    redis_client = None
    print("Redis not available - caching disabled")

# 데이터 모델
class MarketData(BaseModel):
    symbol: str
    prices: List[float]
    volumes: List[float]
    timestamps: List[str]

class PredictionRequest(BaseModel):
    symbol: str
    model: str = "lstm"
    timeframe: str = "1h"
    lookback_period: int = 100

class PredictionResponse(BaseModel):
    symbol: str
    model: str
    current_price: float
    predicted_price: float
    direction: str
    confidence: float
    features: Dict
    timestamp: str

# 모의 LSTM 예측 (실제 모델로 교체 예정)
class MockLSTMPredictor:
    def predict(self, data: MarketData) -> Dict:
        """간단한 모의 예측"""
        prices = data.prices[-100:] if len(data.prices) > 100 else data.prices

        # 간단한 추세 분석
        if len(prices) > 1:
            trend = (prices[-1] - prices[0]) / prices[0]
            volatility = np.std(prices) / np.mean(prices)

            # RSI 계산
            gains = []
            losses = []
            for i in range(1, len(prices)):
                change = prices[i] - prices[i-1]
                if change > 0:
                    gains.append(change)
                    losses.append(0)
                else:
                    gains.append(0)
                    losses.append(abs(change))

            avg_gain = np.mean(gains) if gains else 0
            avg_loss = np.mean(losses) if losses else 0
            rs = avg_gain / avg_loss if avg_loss > 0 else 100
            rsi = 100 - (100 / (1 + rs))

            # 예측
            if rsi > 70:
                direction = "DOWN"
                predicted_change = -0.02
                confidence = 0.7
            elif rsi < 30:
                direction = "UP"
                predicted_change = 0.02
                confidence = 0.7
            else:
                direction = "NEUTRAL"
                predicted_change = trend * 0.1
                confidence = 0.5

            current_price = prices[-1]
            predicted_price = current_price * (1 + predicted_change)

            return {
                "current_price": current_price,
                "predicted_price": predicted_price,
                "direction": direction,
                "confidence": confidence,
                "features": {
                    "rsi": rsi,
                    "volatility": volatility * 100,
                    "trend": trend * 100,
                    "support": min(prices) if prices else 0,
                    "resistance": max(prices) if prices else 0
                }
            }

        return {
            "current_price": prices[-1] if prices else 0,
            "predicted_price": prices[-1] if prices else 0,
            "direction": "NEUTRAL",
            "confidence": 0,
            "features": {}
        }

# 예측 엔진 인스턴스
lstm_predictor = MockLSTMPredictor()

@app.get("/")
async def root():
    return {
        "name": "MONSTA AI Server",
        "version": "1.0.0",
        "status": "running",
        "models": ["lstm", "xgboost", "arima", "ensemble"],
        "timestamp": datetime.now().isoformat()
    }

@app.get("/health")
async def health_check():
    return {
        "status": "healthy",
        "redis": "connected" if redis_client else "disconnected",
        "models_loaded": True,
        "timestamp": datetime.now().isoformat()
    }

@app.post("/predict", response_model=PredictionResponse)
async def predict(request: PredictionRequest):
    """AI 예측 엔드포인트"""

    # 캐시 확인
    cache_key = f"prediction:{request.symbol}:{request.model}"
    if redis_client:
        cached = redis_client.get(cache_key)
        if cached:
            return json.loads(cached)

    # 모의 데이터 생성 (실제로는 DB나 API에서 가져옴)
    current_price = 50000 + np.random.randn() * 1000
    prices = [current_price + np.random.randn() * 500 for _ in range(request.lookback_period)]
    volumes = [1000 + np.random.randn() * 100 for _ in range(request.lookback_period)]

    market_data = MarketData(
        symbol=request.symbol,
        prices=prices,
        volumes=volumes,
        timestamps=[datetime.now().isoformat() for _ in range(request.lookback_period)]
    )

    # 모델별 예측
    if request.model == "lstm":
        result = lstm_predictor.predict(market_data)
    else:
        # 다른 모델들은 추후 구현
        result = lstm_predictor.predict(market_data)

    response = PredictionResponse(
        symbol=request.symbol,
        model=request.model,
        current_price=result["current_price"],
        predicted_price=result["predicted_price"],
        direction=result["direction"],
        confidence=result["confidence"],
        features=result["features"],
        timestamp=datetime.now().isoformat()
    )

    # 캐시 저장 (5초 TTL)
    if redis_client:
        redis_client.setex(
            cache_key,
            5,
            json.dumps(response.dict())
        )

    return response

@app.post("/train")
async def train_model(symbol: str, model: str = "lstm"):
    """모델 학습 엔드포인트"""
    return {
        "status": "training_started",
        "symbol": symbol,
        "model": model,
        "estimated_time": "5 minutes",
        "timestamp": datetime.now().isoformat()
    }

@app.get("/backtest")
async def backtest(symbol: str, model: str = "lstm", days: int = 30):
    """백테스팅 엔드포인트"""

    # 모의 백테스팅 결과
    trades = 100
    wins = 65

    return {
        "symbol": symbol,
        "model": model,
        "period_days": days,
        "total_trades": trades,
        "winning_trades": wins,
        "win_rate": wins / trades * 100,
        "total_return": 15.3,
        "sharpe_ratio": 1.8,
        "max_drawdown": -12.5,
        "avg_profit": 2.4,
        "avg_loss": -1.2,
        "profit_factor": 2.0,
        "timestamp": datetime.now().isoformat()
    }

@app.websocket("/ws")
async def websocket_endpoint(websocket: WebSocket):
    """실시간 예측 WebSocket"""
    await websocket.accept()

    try:
        while True:
            # 클라이언트 메시지 수신
            data = await websocket.receive_text()
            request = json.loads(data)

            # 예측 수행
            if request.get("action") == "predict":
                symbol = request.get("symbol", "BTCUSDT")

                # 모의 예측 데이터
                prediction = {
                    "type": "prediction",
                    "symbol": symbol,
                    "price": 50000 + np.random.randn() * 1000,
                    "direction": np.random.choice(["UP", "DOWN", "NEUTRAL"]),
                    "confidence": np.random.random(),
                    "timestamp": datetime.now().isoformat()
                }

                await websocket.send_text(json.dumps(prediction))

            # 1초 대기
            await asyncio.sleep(1)

    except Exception as e:
        print(f"WebSocket error: {e}")
    finally:
        await websocket.close()

@app.get("/models")
async def list_models():
    """사용 가능한 모델 목록"""
    return {
        "models": [
            {
                "name": "lstm",
                "type": "Deep Learning",
                "accuracy": 0.72,
                "description": "Long Short-Term Memory 신경망"
            },
            {
                "name": "xgboost",
                "type": "Gradient Boosting",
                "accuracy": 0.68,
                "description": "Extreme Gradient Boosting"
            },
            {
                "name": "arima",
                "type": "Time Series",
                "accuracy": 0.65,
                "description": "AutoRegressive Integrated Moving Average"
            },
            {
                "name": "ensemble",
                "type": "Ensemble",
                "accuracy": 0.75,
                "description": "모든 모델의 앙상블"
            }
        ]
    }

@app.get("/performance")
async def get_performance():
    """시스템 성능 메트릭"""
    return {
        "predictions_per_second": 1000,
        "cache_hit_rate": 0.85,
        "avg_latency_ms": 50,
        "models_loaded": 4,
        "memory_usage_mb": 512,
        "cpu_usage_percent": 25,
        "timestamp": datetime.now().isoformat()
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(
        "main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="info"
    )