python-ai-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tensorflow==2.14.0
torch==2.1.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
ta==0.10.2
yfinance==0.2.32
redis==5.0.1
pydantic==2.5.0
python-dotenv==1.0.0
websocket-client==1.6.4
ccxt==4.1.22
statsmodels==0.14.0